Fix code, update import paths based on move
