+++ src/dodal/beamlines/i20_1.py
@@ -1,5 +1,5 @@
-from dodal.beamlines.beamline_utils import device_instantiation
-from dodal.beamlines.beamline_utils import set_beamline as set_utils_beamline
+from dodal.beamlines_common.beamline_utils import device_instantiation
+from dodal.beamlines_common.beamline_utils import set_beamline as set_utils_beamline
 from dodal.devices.turbo_slit import TurboSlit
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import get_beamline_name
