+++ src/dodal/beamlines/i22.py
@@ -3,12 +3,13 @@ from pathlib import Path
 from ophyd_async.epics.areadetector import AravisDetector, PilatusDetector
 from ophyd_async.panda import HDFPanda
 
-from dodal.beamlines.beamline_utils import (
+from dodal.beamlines_common.beamline_utils import (
     device_instantiation,
     get_directory_provider,
     set_directory_provider,
 )
-from dodal.beamlines.beamline_utils import set_beamline as set_utils_beamline
+from dodal.beamlines_common.beamline_utils import set_beamline as set_utils_beamline
+from dodal.beamlines_common.device_helpers import numbered_slits
 from dodal.common.visit import LocalDirectoryServiceClient, StaticVisitDirectoryProvider
 from dodal.devices.focusing_mirror import FocusingMirror
 from dodal.devices.i22.fswitch import FSwitch
@@ -16,10 +17,6 @@ from dodal.devices.slits import Slits
 from dodal.devices.tetramm import TetrammDetector
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import get_beamline_name, skip_device
-
-from ._device_helpers import numbered_slits
-from .beamline_utils import device_instantiation, get_directory_provider
-from .beamline_utils import set_beamline as set_utils_beamline
 
 BL = get_beamline_name("i22")
 set_log_beamline(BL)
