+++ src/dodal/beamlines/i24.py
@@ -1,5 +1,5 @@
-from dodal.beamlines.beamline_utils import BL, device_instantiation
-from dodal.beamlines.beamline_utils import set_beamline as set_utils_beamline
+from dodal.beamlines_common.beamline_utils import BL, device_instantiation
+from dodal.beamlines_common.beamline_utils import set_beamline as set_utils_beamline
 from dodal.devices.detector import DetectorParams
 from dodal.devices.eiger import EigerDetector
 from dodal.devices.i24.dual_backlight import DualBacklight
