+++ tests/beamlines/unit_tests/test_beamline_utils.py
@@ -8,7 +8,8 @@ from ophyd.sim import FakeEpicsSignal
 from ophyd_async.core import Device as OphydV2Device
 from ophyd_async.core import StandardReadable
 
-from dodal.beamlines import beamline_utils, i03
+from dodal.beamlines import i03
+from dodal.beamlines_common import beamline_utils
 from dodal.devices.aperturescatterguard import ApertureScatterguard
 from dodal.devices.smargon import Smargon
 from dodal.devices.zebra import Zebra
