+++ tests/beamlines/unit_tests/test_device_instantiation.py
@@ -2,7 +2,7 @@ from typing import Any
 
 import pytest
 
-from dodal.beamlines import beamline_utils
+from dodal.beamlines_common import beamline_utils
 from dodal.utils import BLUESKY_PROTOCOLS, make_all_devices
 
 ALL_BEAMLINES = {"i03", "i04", "i04_1", "i20_1", "i22", "i23", "i24", "p38", "p45"}
